+++ src/Api1/Diogo.Softplan.Challenge.Api1.Domain/Services/ITaxaDeJurosService.cs
@@ -2,6 +2,6 @@
 {
     public interface ITaxaDeJurosService
     {
-        double Obter();
+        decimal Obter();
     }
 }
